refactor(admin): colocate IMS connections management module
Connections page, connection CRUD logic and styles were scattered across generic folders. Move Connections.vue, useConnections.ts and connections.scss under admin/connections so the IMS connections feature is easier to maintain.
