+++ webui/src/admin/connections/connections.scss
